backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
scipy==1.14.1
matplotlib==3.9.2
seaborn==0.13.2
requests==2.32.3
aiofiles==24.1.0
redis==5.1.1
celery==5.4.0
Pillow==10.4.0
